Move HadoopPathUtils to parquet module
