+++ java/ingest/ingest-runner/src/main/java/sleeper/ingest/job/IngestJobRunner.java
@@ -37,6 +37,7 @@ import sleeper.io.parquet.record.ParquetReaderIterator;
 import sleeper.io.parquet.record.ParquetRecordReader;
 import sleeper.statestore.StateStoreException;
 import sleeper.statestore.StateStoreProvider;
+import sleeper.utils.HadoopPathUtils;
 
 import java.io.IOException;
 import java.util.ArrayList;
